remove octoprint script + rework setup + scrollable custom actions
