+++ package.json
@@ -86,7 +86,6 @@
     "compare-versions": "^3.6.0",
     "electron-store": "^6.0.1",
     "got": "^11.8.0",
-    "jquery": "^3.5.1",
     "lodash": "^4.17.20",
     "ngx-electron": "^2.2.0",
     "ngx-spinner": "^10.0.1",
