+++ src/app/config/config.service.ts
@@ -4,6 +4,8 @@ import Ajv from 'ajv';
 import _ from 'lodash';
 
 import { NotificationService } from '../notification/notification.service';
+import { Config, CustomAction, HttpHeader, URLSplit } from './config.model';
+import { configSchema } from './config.schema';
 
 @Injectable({
   providedIn: 'root',
@@ -22,7 +24,7 @@ export class ConfigService {
 
   public constructor(private notificationService: NotificationService) {
     const ajv = new Ajv({ allErrors: true });
-    this.validator = ajv.compile(schema);
+    this.validator = ajv.compile(configSchema);
     try {
       const Store = window.require('electron-store');
       this.store = new Store();
@@ -40,16 +42,20 @@ export class ConfigService {
     this.config = config;
     this.valid = this.validate();
     if (this.valid) {
-      this.httpHeaders = {
-        headers: new HttpHeaders({
-          'x-api-key': this.config.octoprint.accessToken,
-          'Cache-Control': 'no-cache',
-          Pragma: 'no-cache',
-          Expires: '0',
-        }),
-      };
+      this.generateHttpHeaders();
     }
     this.initialized = true;
+  }
+
+  public generateHttpHeaders(): void {
+    this.httpHeaders = {
+      headers: new HttpHeaders({
+        'x-api-key': this.config.octoprint.accessToken,
+        'Cache-Control': 'no-cache',
+        Pragma: 'no-cache',
+        Expires: '0',
+      }),
+    };
   }
 
   public getRemoteConfig(): Config {
@@ -85,48 +91,44 @@ export class ConfigService {
     return errors;
   }
 
-  public saveConfig(config?: Config): string {
-    if (!config) {
-      config = this.config;
-    }
-    if (window && window.process && window.process.type) {
+  public saveConfig(config: Config): string {
+    try {
       this.store.set('config', config);
       const configStored = this.store.get('config');
       if (this.validateGiven(configStored)) {
+        this.config = config;
+        this.generateHttpHeaders();
         return null;
       } else {
         return 'Saved config is invalid!';
       }
-    } else {
-      return "Browser version doesn't support saving!";
+    } catch {
+      return 'Saving config failed!';
     }
   }
 
-  public updateConfig(): void {
-    if (window && window.process && window.process.type) {
-      this.update = false;
-      this.initialize(this.store.get('config'));
-    }
-  }
+  public splitOctoprintURL(octoprintURL: string): URLSplit {
+    const host = octoprintURL.split(':')[1].replace('//', '');
+    const port = parseInt(octoprintURL.split(':')[2].replace('/api/', ''), 10);
 
-  public revertConfigForInput(config: Config): Config {
-    config.octoprint.urlSplit = {
-      url: config.octoprint.url.split(':')[1].replace('//', ''),
-      port: parseInt(config.octoprint.url.split(':')[2].replace('/api/', ''), 10),
+    return {
+      host,
+      port: isNaN(port) ? null : port,
     };
-    if (isNaN(config.octoprint.urlSplit.port)) {
-      config.octoprint.urlSplit.port = null;
+  }
+
+  public mergeOctoprintURL(urlSplit: URLSplit): string {
+    // TODO: remove api/ from URL for v2.2.0
+    if (urlSplit.port !== null || !isNaN(urlSplit.port)) {
+      return `http://${urlSplit.host}:${urlSplit.port}/api/`;
+    } else {
+      return `http://${urlSplit.host}/api/`;
     }
-    return config;
   }
 
   public createConfigFromInput(config: Config): Config {
     const configOut = _.cloneDeep(config);
-    if (config.octoprint.urlSplit.port !== null || !isNaN(config.octoprint.urlSplit.port)) {
-      configOut.octoprint.url = `http://${config.octoprint.urlSplit.url}:${config.octoprint.urlSplit.port}/api/`;
-    } else {
-      configOut.octoprint.url = `http://${config.octoprint.urlSplit.url}/api/`;
-    }
+    configOut.octoprint.url = this.mergeOctoprintURL(config.octoprint.urlSplit);
     delete configOut.octoprint.urlSplit;
     return configOut;
   }
@@ -267,385 +269,3 @@ export class ConfigService {
     return this.config.octodash.previewProgressCircle;
   }
 }
-
-interface HttpHeader {
-  headers: HttpHeaders;
-}
-
-export interface Config {
-  octoprint: Octoprint;
-  printer: Printer;
-  filament: Filament;
-  plugins: Plugins;
-  octodash: OctoDash;
-}
-
-interface Octoprint {
-  url: string;
-  accessToken: string;
-  urlSplit?: {
-    url: string;
-    port: number;
-  };
-}
-
-interface Printer {
-  name: string;
-  xySpeed: number;
-  zSpeed: number;
-  zBabystepGCode: string;
-  defaultTemperatureFanSpeed: DefaultTemperatureFanSpeed;
-}
-
-interface DefaultTemperatureFanSpeed {
-  hotend: number;
-  heatbed: number;
-  fan: number;
-}
-
-interface Filament {
-  thickness: number;
-  density: number;
-  feedLength: number;
-  feedSpeed: number;
-  feedSpeedSlow: number;
-  purgeDistance: number;
-  useM600: boolean;
-}
-
-interface Plugins {
-  displayLayerProgress: Plugin;
-  enclosure: EnclosurePlugin;
-  filamentManager: Plugin;
-  preheatButton: Plugin;
-  printTimeGenius: Plugin;
-  psuControl: PSUControlPlugin;
-}
-
-interface Plugin {
-  enabled: boolean;
-}
-
-interface EnclosurePlugin extends Plugin {
-  ambientSensorID: number | null;
-  filament1SensorID: number | null;
-  filament2SensorID: number | null;
-}
-
-interface PSUControlPlugin extends Plugin {
-  turnOnPSUWhenExitingSleep: boolean;
-}
-
-interface OctoDash {
-  customActions: CustomAction[];
-  fileSorting: FileSorting;
-  pollingInterval: number;
-  touchscreen: boolean;
-  turnScreenOffWhileSleeping: boolean;
-  preferPreviewWhilePrinting: boolean;
-  previewProgressCircle: boolean;
-}
-
-interface CustomAction {
-  icon: string;
-  command: string;
-  color: string;
-  confirm: boolean;
-  exit: boolean;
-}
-
-interface FileSorting {
-  attribute: 'name' | 'date' | 'size';
-  order: 'asc' | 'dsc';
-}
-
-const schema = {
-  definitions: {},
-  $schema: 'http://json-schema.org/draft-07/schema#',
-  $id: 'http://example.com/root.json',
-  type: 'object',
-  required: ['octoprint', 'printer', 'filament', 'plugins', 'octodash'],
-  properties: {
-    octoprint: {
-      $id: '#/properties/octoprint',
-      type: 'object',
-      required: ['accessToken', 'url'],
-      properties: {
-        accessToken: {
-          $id: '#/properties/octoprint/properties/accessToken',
-          type: 'string',
-          pattern: '^(.*)$',
-        },
-        url: {
-          $id: '#/properties/octoprint/properties/url',
-          type: 'string',
-          pattern: '^(.*)$',
-        },
-      },
-    },
-    printer: {
-      $id: '#/properties/printer',
-      type: 'object',
-      required: ['name', 'xySpeed', 'zSpeed', 'zBabystepGCode', 'defaultTemperatureFanSpeed'],
-      properties: {
-        name: {
-          $id: '#/properties/printer/properties/name',
-          type: 'string',
-          pattern: '^(.*)$',
-        },
-        xySpeed: {
-          $id: '#/properties/printer/properties/xySpeed',
-          type: 'integer',
-        },
-        zSpeed: {
-          $id: '#/properties/printer/properties/zSpeed',
-          type: 'integer',
-        },
-        zBabystepGCode: {
-          $id: '#/properties/printer/properties/zBabystepGCode',
-          type: 'string',
-          pattern: '^(.*)$',
-        },
-        defaultTemperatureFanSpeed: {
-          $id: '#/properties/printer/properties/defaultTemperatureFanSpeed',
-          type: 'object',
-          required: ['hotend', 'heatbed', 'fan'],
-          properties: {
-            hotend: {
-              $id: '#/properties/printer/properties/defaultTemperatureFanSpeed/hotend',
-              type: 'integer',
-            },
-            heatbed: {
-              $id: '#/properties/printer/properties/defaultTemperatureFanSpeed/heatbed',
-              type: 'integer',
-            },
-            fan: {
-              $id: '#/properties/printer/properties/defaultTemperatureFanSpeed/fan',
-              type: 'integer',
-            },
-          },
-        },
-      },
-    },
-    filament: {
-      $id: '#/properties/filament',
-      type: 'object',
-      required: ['density', 'thickness', 'feedLength', 'feedSpeed', 'feedSpeedSlow', 'purgeDistance', 'useM600'],
-      properties: {
-        density: {
-          $id: '#/properties/filament/properties/density',
-          type: 'number',
-        },
-        thickness: {
-          $id: '#/properties/filament/properties/thickness',
-          type: 'number',
-        },
-        feedLength: {
-          $id: '#/properties/filament/properties/feedLength',
-          type: 'integer',
-        },
-        feedSpeed: {
-          $id: '#/properties/filament/properties/feedSpeed',
-          type: 'number',
-        },
-        feedSpeedSlow: {
-          $id: '#/properties/filament/properties/feedSpeedSlow',
-          type: 'number',
-        },
-        purgeDistance: {
-          $id: '#/properties/filament/properties/purgeDistance',
-          type: 'integer',
-        },
-        useM600: {
-          $id: '#/properties/filament/properties/useM600',
-          type: 'boolean',
-        },
-      },
-    },
-    plugins: {
-      $id: '#/properties/plugins',
-      type: 'object',
-      required: [
-        'displayLayerProgress',
-        'enclosure',
-        'filamentManager',
-        'preheatButton',
-        'printTimeGenius',
-        'psuControl',
-      ],
-      properties: {
-        displayLayerProgress: {
-          $id: '#/properties/plugins/properties/displayLayerProgress',
-          type: 'object',
-          required: ['enabled'],
-          properties: {
-            enabled: {
-              $id: '#/properties/plugins/properties/displayLayerProgress/properties/enabled',
-              type: 'boolean',
-            },
-          },
-        },
-        enclosure: {
-          $id: '#/properties/plugins/properties/enclosure',
-          type: 'object',
-          required: ['enabled', 'ambientSensorID', 'filament1SensorID', 'filament2SensorID'],
-          properties: {
-            enabled: {
-              $id: '#/properties/plugins/properties/enclosure/properties/enabled',
-              type: 'boolean',
-            },
-            ambientSensorID: {
-              $id: '#/properties/plugins/properties/enclosure/properties/ambientSensorID',
-              type: ['number', 'null'],
-              pattern: '^(.*)$',
-            },
-            filament1SensorID: {
-              $id: '#/properties/plugins/properties/enclosure/properties/filament1SensorID',
-              type: ['number', 'null'],
-              pattern: '^(.*)$',
-            },
-            filament2SensorID: {
-              $id: '#/properties/plugins/properties/enclosure/properties/filament2SensorID',
-              type: ['number', 'null'],
-              pattern: '^(.*)$',
-            },
-          },
-        },
-        filamentManager: {
-          $id: '#/properties/plugins/properties/filamentManager',
-          type: 'object',
-          required: ['enabled'],
-          properties: {
-            enabled: {
-              $id: '#/properties/plugins/properties/filamentManager/properties/enabled',
-              type: 'boolean',
-            },
-          },
-        },
-        preheatButton: {
-          $id: '#/properties/plugins/properties/preheatButton',
-          type: 'object',
-          required: ['enabled'],
-          properties: {
-            enabled: {
-              $id: '#/properties/plugins/properties/preheatButton/properties/enabled',
-              type: 'boolean',
-            },
-          },
-        },
-        printTimeGenius: {
-          $id: '#/properties/plugins/properties/printTimeGenius',
-          type: 'object',
-          required: ['enabled'],
-          properties: {
-            enabled: {
-              $id: '#/properties/plugins/properties/printTimeGenius/properties/enabled',
-              type: 'boolean',
-            },
-          },
-        },
-        psuControl: {
-          $id: '#/properties/plugins/properties/psuControl',
-          type: 'object',
-          required: ['enabled', 'turnOnPSUWhenExitingSleep'],
-          properties: {
-            enabled: {
-              $id: '#/properties/plugins/properties/printTimeGenius/properties/enabled',
-              type: 'boolean',
-            },
-            turnOnPSUWhenExitingSleep: {
-              $id: '#/properties/plugins/properties/turnOnPSUWhenExitingSleep',
-              type: 'boolean',
-            },
-          },
-        },
-      },
-    },
-    octodash: {
-      $id: '#/properties/octodash',
-      type: 'object',
-      required: [
-        'customActions',
-        'fileSorting',
-        'pollingInterval',
-        'touchscreen',
-        'turnScreenOffWhileSleeping',
-        'preferPreviewWhilePrinting',
-        'previewProgressCircle',
-      ],
-      properties: {
-        customActions: {
-          $id: '#/properties/octodash/properties/customActions',
-          type: 'array',
-          items: {
-            $id: '#/properties/octodash/properties/customActions/items',
-            type: 'object',
-            required: ['icon', 'command', 'color', 'confirm', 'exit'],
-            properties: {
-              icon: {
-                $id: '#/properties/octodash/properties/customActions/items/properties/icon',
-                type: 'string',
-                pattern: '^(.*)$',
-              },
-              command: {
-                $id: '#/properties/octodash/properties/customActions/items/properties/command',
-                type: 'string',
-                pattern: '^(.*)$',
-              },
-              color: {
-                $id: '#/properties/octodash/properties/customActions/items/properties/color',
-                type: 'string',
-                pattern: '^(.*)$',
-              },
-              confirm: {
-                $id: '#/properties/octodash/properties/customActions/items/properties/confirm',
-                type: 'boolean',
-              },
-              exit: {
-                $id: '#/properties/octodash/properties/customActions/items/properties/exit',
-                type: 'boolean',
-              },
-            },
-          },
-        },
-        fileSorting: {
-          $id: '#/properties/octodash/properties/fileSorting',
-          type: 'object',
-          required: ['attribute', 'order'],
-          properties: {
-            attribute: {
-              $id: '#/properties/octodash/properties/fileSorting/properties/attribute',
-              type: 'string',
-              pattern: '^(name|date|size)$',
-            },
-            order: {
-              $id: '#/properties/octodash/properties/fileSorting/properties/order',
-              type: 'string',
-              pattern: '^(asc|dsc)$',
-            },
-          },
-        },
-        pollingInterval: {
-          $id: '#/properties/octodash/properties/pollingInterval',
-          type: 'integer',
-        },
-        touchscreen: {
-          $id: '#/properties/octodash/properties/touchscreen',
-          type: 'boolean',
-        },
-        turnScreenOffWhileSleeping: {
-          $id: '#/properties/octodash/properties/turnScreenOffWhileSleeping',
-          type: 'boolean',
-        },
-        preferPreviewWhilePrinting: {
-          $id: '#/properties/octodash/properties/preferPreviewWhilePrinting',
-          type: 'boolean',
-        },
-        previewProgressCircle: {
-          $id: '#/properties/octodash/properties/previewProgressCircle',
-          type: 'boolean',
-        },
-      },
-    },
-  },
-};
