+++ src/app/config/invalid/invalid.component.scss
