+++ src/app/config/setup/extruder-information/extruder-information.component.scss
@@ -1,0 +1,7 @@
+form {
+  margin-left: 4vw;
+}
+
+.setup__explanation {
+  margin-top: 4vh !important;
+}
