+++ src/app/config/setup/octoprint-authentication/octoprint-authentication.component.html
@@ -1,0 +1,20 @@
+<span class="setup__text">
+  Please authenticate me now. You can either click the button below and confirm the request in your OctoPrint
+  webinterface or enter an API Key manually.
+</span>
+<div class="octoprint-authentication__login-button">
+  <span class="setup__button" (click)="loginWithOctoprintUI()">send request</span>
+</div>
+<form class="setup__form">
+  <label for="accessToken" class="setup__input-label">API Key:</label>
+  <input
+    type="text"
+    id="accessToken"
+    class="setup__input"
+    [ngModel]="this.accessToken"
+    (ngModelChange)="accessTokenChange.emit($event)"
+    name="accessToken"
+    style="width: 67vw"
+    required
+  />
+</form>
