+++ src/app/config/setup/octoprint-authentication/octoprint-authentication.component.scss
@@ -1,0 +1,6 @@
+.octoprint-authentication {
+  &__login-button {
+    text-align: center;
+    padding: 8vh 0 0;
+  }
+}
