+++ src/app/config/setup/personalization/personalization.component.scss
@@ -1,0 +1,5 @@
+.personalization {
+  &__input {
+    margin-top: 5vh;
+  }
+}
