+++ src/app/config/setup/welcome/welcome.component.scss
@@ -1,0 +1,10 @@
+.welcome-screen {
+  &__heading {
+    display: block;
+    text-align: center;
+    font-size: 6vw;
+    margin-top: 2vh;
+    margin-bottom: 6vh;
+    font-weight: 500;
+  }
+}
