+++ src/app/settings/settings.component.html
@@ -53,7 +53,7 @@
               class="settings__input"
               name="octoprintURLName"
               style="width: 23.2vw"
-              [(ngModel)]="this.config.octoprint.urlSplit.url"
+              [(ngModel)]="this.config.octoprint.urlSplit.host"
               required
             /><span class="settings__input-divider">:</span>
             <input
